clinique/requirements.txt
asgiref==3.8.1
Django==5.2.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
PyJWT==2.9.0
sqlparse==0.5.3
tzdata==2025.2
python-dotenv==1.0.0
django-cors-headers==4.3.1
gunicorn==21.2.0

# Development/Quality
black==24.1.1
flake8==7.0.0
isort==5.13.2
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
pre-commit==3.6.0

# API Documentation
drf-spectacular==0.27.0
drf-spectacular[sidecar]==0.27.0

# Logging & Monitoring
python-json-logger==2.0.7
